pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp_amazon_affiliate"
version = "0.1.0"
description = "MCP Amazon Affiliate Integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "python-amazon-paapi>=5.0.1",
    "agno>=1.5.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
mcp-amazon-affiliate = "mcp_amazon_affiliate:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "server.py",
    "README.md",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
